requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alibabacloud-credentials==0.2.0
alibabacloud-dysmsapi20170525==2.0.9
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-openapi-util==0.1.6
alibabacloud-tea==0.2.9
alibabacloud-tea-openapi==0.3.3
alibabacloud-tea-util==0.3.5
alibabacloud-tea-xml==0.0.2
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.15.0
amqp==5.1.1
arrow==1.2.2
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.6.0
billiard==3.6.4.0
binaryornot==0.4.4
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
configparser==5.2.0
constantly==15.1.0
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==36.0.2
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.7
django-channels==0.7.0
django-crispy-forms==1.12.0
django-filter==21.1
django-formtools==2.3
django-import-export==2.6.0
django-minio-storage==0.3.10
django-oss-storage==1.1.1
django-redis==5.2.0
django-reversion==4.0.0
django-simpleui==2022.3.19
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
et-xmlfile==1.1.0
eventlet==0.33.1
filetype==1.0.13
flower==1.1.0
frozenlist==1.3.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.19.1
humanize==4.2.3
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
install==1.3.4
itypes==1.2.0
Jinja2==3.1.1
jinja2-time==0.2.0
jmespath==0.10.0
kombu==5.2.4
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==2.1.1
minio==6.0.2
multidict==6.0.2
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
oss2==2.15.0
packaging==21.3
pika==1.2.0
Pillow==9.1.0
poyo==0.5.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.11.2
PyJWT==2.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2022.1
PyYAML==5.4.1
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
supervisor==4.2.4
tablib==3.0.0
text-unidecode==1.3
toml==0.10.2
tornado==6.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.9
uuid==1.30
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.3
whitenoise==6.0.0
wrapt==1.14.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
